refactor(handlers): move consent body generation to the oidc client
