+++ internal/handlers/oidc.go
@@ -23,30 +23,6 @@ func isConsentMissing(workflow *session.OIDCWorkflowSession, requestedScopes, re
 		len(requestedAudience) > 0 && utils.IsStringSlicesDifferentFold(requestedAudience, workflow.GrantedAudience)
 }
 
-func scopeNamesToScopes(scopeSlice []string) (scopes []Scope) {
-	for _, name := range scopeSlice {
-		if val, ok := scopeDescriptions[name]; ok {
-			scopes = append(scopes, Scope{name, val})
-		} else {
-			scopes = append(scopes, Scope{name, name})
-		}
-	}
-
-	return scopes
-}
-
-func audienceNamesToAudience(scopeSlice []string) (audience []Audience) {
-	for _, name := range scopeSlice {
-		if val, ok := audienceDescriptions[name]; ok {
-			audience = append(audience, Audience{name, val})
-		} else {
-			audience = append(audience, Audience{name, name})
-		}
-	}
-
-	return audience
-}
-
 func newOIDCSession(ctx *middlewares.AutheliaCtx, ar fosite.AuthorizeRequester) (session *openid.DefaultSession, err error) {
 	userSession := ctx.GetSession()
