+++ internal/handlers/register_oidc.go
@@ -9,7 +9,7 @@ import (
 // RegisterOIDC registers the handlers with the fasthttp *router.Router. TODO: Add paths for UserInfo, Flush, Logout.
 func RegisterOIDC(router *router.Router, middleware middlewares.RequestHandlerBridge) {
 	// TODO: Add OPTIONS handler.
-	router.GET(oidcWellKnownPath, middleware(oidcWellKnown))
+	router.GET("/.well-known/openid-configuration", middleware(oidcWellKnown))
 
 	router.GET(oidcConsentPath, middleware(oidcConsent))
