+++ internal/handlers/types_oidc.go
@@ -1,9 +1,5 @@
 package handlers
 
-import (
-	"github.com/dgrijalva/jwt-go"
-)
-
 // ConsentPostRequestBody schema of the request body of the consent POST endpoint.
 type ConsentPostRequestBody struct {
 	ClientID       string `json:"client_id"`
@@ -13,35 +9,6 @@ type ConsentPostRequestBody struct {
 // ConsentPostResponseBody schema of the response body of the consent POST endpoint.
 type ConsentPostResponseBody struct {
 	RedirectURI string `json:"redirect_uri"`
-}
-
-// ConsentGetResponseBody schema of the response body of the consent GET endpoint.
-type ConsentGetResponseBody struct {
-	ClientID          string     `json:"client_id"`
-	ClientDescription string     `json:"client_description"`
-	Scopes            []Scope    `json:"scopes"`
-	Audience          []Audience `json:"audience"`
-}
-
-// Scope represents the scope information.
-type Scope struct {
-	Name        string `json:"name"`
-	Description string `json:"description"`
-}
-
-// Audience represents the audience information.
-type Audience struct {
-	Name        string `json:"name"`
-	Description string `json:"description"`
-}
-
-// OIDCClaims represents a set of OIDC claims.
-type OIDCClaims struct {
-	jwt.StandardClaims
-
-	Workflow        string   `json:"workflow"`
-	Username        string   `json:"username,omitempty"`
-	RequestedScopes []string `json:"requested_scopes,omitempty"`
 }
 
 // WellKnownConfigurationJSON is the OIDC well known config struct.
