+++ internal/oidc/client.go
@@ -7,6 +7,7 @@ import (
 	"github.com/authelia/authelia/internal/authentication"
 	"github.com/authelia/authelia/internal/authorization"
 	"github.com/authelia/authelia/internal/configuration/schema"
+	"github.com/authelia/authelia/internal/session"
 )
 
 // NewClient creates a new InternalClient.
@@ -62,6 +63,17 @@ func (c InternalClient) IsAuthenticationLevelSufficient(level authentication.Lev
 // GetID returns the ID.
 func (c InternalClient) GetID() string {
 	return c.ID
+}
+
+func (c InternalClient) GetConsentRequestBody(session *session.OIDCWorkflowSession) (body ConsentGetResponseBody) {
+	body.ClientID = c.ID
+	body.ClientDescription = c.Description
+	if session != nil {
+		body.Scopes = scopeNamesToScopes(session.RequestedScopes)
+		body.Audience = audienceNamesToAudience(session.RequestedAudience)
+	}
+
+	return body
 }
 
 // GetHashedSecret returns the Secret.
