+++ internal/oidc/const.go
@@ -1,0 +1,10 @@
+package oidc
+
+var scopeDescriptions = map[string]string{
+	"openid":  "Use OpenID to verify your identity",
+	"email":   "Access your email addresses",
+	"profile": "Access your username",
+	"groups":  "Access your group membership",
+}
+
+var audienceDescriptions = map[string]string{}
