+++ internal/oidc/types.go
@@ -1,0 +1,21 @@
+package oidc
+
+// ConsentGetResponseBody schema of the response body of the consent GET endpoint.
+type ConsentGetResponseBody struct {
+	ClientID          string     `json:"client_id"`
+	ClientDescription string     `json:"client_description"`
+	Scopes            []Scope    `json:"scopes"`
+	Audience          []Audience `json:"audience"`
+}
+
+// Scope represents the scope information.
+type Scope struct {
+	Name        string `json:"name"`
+	Description string `json:"description"`
+}
+
+// Audience represents the audience information.
+type Audience struct {
+	Name        string `json:"name"`
+	Description string `json:"description"`
+}
